Move AspNetCache and related items to new namespace: SenseNet.ContentRepository.Storage.Caching.Legacy
