--- src/Storage/Caching/ICache.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Web;
-
-namespace SenseNet.ContentRepository.Storage.Caching
-{
-    public interface ICache : ISnCache
-    {
-        DateTime NoAbsoluteExpiration { get; }
-        TimeSpan NoSlidingExpiration { get; }
-
-        long EffectivePercentagePhysicalMemoryLimit { get; }
-        long EffectivePrivateBytesLimit { get; }
-
-        HttpContext CurrentHttpContext { get; set; }
-    }
-}
