+++ src/Storage/SenseNet.Storage.csproj
@@ -205,14 +205,14 @@
     <Compile Include="AppModel\RepositoryEventHandlerBase.cs" />
     <Compile Include="AppModel\RepositoryEventRouter.cs" />
     <Compile Include="BinaryData.cs" />
-    <Compile Include="Caching\AspNetCache.cs" />
-    <Compile Include="Caching\CacheBase.cs" />
+    <Compile Include="Caching\Legacy\AspNetCache.cs" />
+    <Compile Include="Caching\Legacy\CacheBase.cs" />
     <Compile Include="Caching\Dependency\CacheDependencyFactory.cs" />
-    <Compile Include="Caching\Dependency\NodeTypeDependency.cs" />
+    <Compile Include="Caching\Legacy\NodeTypeDependencyImpl.cs" />
     <Compile Include="Caching\Dependency\EventArgsGeneric.cs" />
-    <Compile Include="Caching\Dependency\NodeIdDependency.cs" />
-    <Compile Include="Caching\Dependency\PathDependency.cs" />
-    <Compile Include="Caching\ICache.cs" />
+    <Compile Include="Caching\Legacy\NodeIdDependencyImpl.cs" />
+    <Compile Include="Caching\Legacy\PathDependencyImpl.cs" />
+    <Compile Include="Caching\Legacy\ICache.cs" />
     <Compile Include="Diagnostics\TimeboxedActivity.cs" />
     <Compile Include="DistributedApplication\Messaging\MsmqChannel.cs" />
     <Compile Include="DataBackingStore.cs" />
@@ -258,7 +258,7 @@
     <Compile Include="Caching\DistributedActions\CleanupCacheAction.cs" />
     <Compile Include="Caching\DistributedActions\PortletChangedAction.cs" />
     <Compile Include="Caching\DistributedActions\PortletChangedMessage.cs" />
-    <Compile Include="Caching\Dependency\PortletDependency.cs" />
+    <Compile Include="Caching\Legacy\PortletDependencyImpl.cs" />
     <Compile Include="DistributedApplication.cs" />
     <Compile Include="DistributedApplication\Messaging\BinaryMessageFormatter.cs" />
     <Compile Include="DistributedApplication\Messaging\ClusterChannel.cs" />
